Move CA secret code to a separate file
* pkg/operator/controller/controller.go (caCertSecretName):
(ensureRouterCACertificateSecret, getRouterCA): Move from here...
* pkg/operator/controller/controller_ca_secret.go: ...to this new file.
